requirements.txt
aiocache==0.11.1
aiocqhttp==1.2.5
aiofiles==0.5.0
APScheduler==3.6.3
beautifulsoup4==4.9.1
blinker==1.4
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.20
DBUtils==1.3
fuzzywuzzy==0.18.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.6.16
httpcore==0.9.1
httpx==0.13.3
Hypercorn==0.10.1
hyperframe==5.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
nonebot==1.6.0
numpy==1.19.0
Pillow==7.1.2
priority==1.3.0
pymssql==2.1.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-lzf
pytz==2020.1
Quart==0.11.5
requests==2.24.0
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
toml==0.10.1
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
wsproto==0.15.0
xlrd==1.2.0
